+++ anchor-image-bean/src/main/java/org/anchoranalysis/image/bean/threshold/relation/BinaryOnVoxels.java
@@ -43,7 +43,7 @@ public class BinaryOnVoxels extends BinaryVoxelsBase {
 
     @Override
     public double threshold() {
-        return (double) (BinaryValues.getDefault().getOnInt() - 1);
+        return BinaryValues.getDefault().getOnInt() - 1.0;
     }
 
     @Override
